CMake: Honor CC and CXX environment variable
Do not force a specific c(++) compiler by setting the
CMAKE_{}_COMPILER variable.
After the c(++) compiler binary has been choosen check if it is the GNU compilear.

In this way we allow the users to specify a GNU compiler under a non
standard path through the CC and CXX envirnoment variable.