docker: added image file for build-env container
The new docker folder will host all the docker image files.
At the moment there is only one image file that allows to build a container based on CentOS 7 that provides a complete build environment for ADAFS. It is bundled with all the required dependencies, both from system repository and the ones that need to be compiled from source. It also provide a suitable compiler (gcc 7.2.1) for later use.

This container can be use as developing environment as well as the base for other images. For instance we can use it to build a container that provide the ADAFS daemon.