cmake: remove redundant compile flag declarations
The preload specific cmake files inherits the global flag declarations,
thus it is not necessary to specify again all the flags.
Moreover the only addition made by this declaration is the `-fPIC` flag.
According to the official CMake docs [1], is set automatically for
library target as SHARED. Thus we don't need to set it explicitely.

[1]: https://cmake.org/cmake/help/v3.6/prop_tgt/POSITION_INDEPENDENT_CODE.html