Daemon now creates a pid file with pid and mountdir info for clients
Previously, clients would look for the daemon by name which was not robust.
If the daemon was started with valgrind for example, clients would not find
the daemon.

Similar to other applications a daemon now creates a pid file with
information about its pid and its started mountdir.

Proper errorhandling on server and client side was added.

Multiple daemons per node are explicitly not allowed for now!