Preload: RPC addresses of hosts are now obtained at startup
Previously, host rpc addresses have been looked up lazily when a request
was done. This is not necessary as all hosts are looked up anyways
while the file system is running.

Now, all RPC addresses are looked up when the environment is initialized.
After that, the rpc_address map is only accessed read-only and does not
require a mutex anymore.