dl script: exit on donwload task errors
Problem: If one of the donwload task fails the script do not stop and
the other task continue to donwload sources.

Solution: If one of the task fails, print an informative error message
and send a signal to the parent process that kills all the spawned tasks
and exits with error code.