deprecated *stat64 functions family
According to the glibc [1] the stat64 function family would be called only
on 32-bit machine.

To be sure we still intercept the function, but we now throw a NOTSUP
error.

In this way we can drop the adafs_stat64 function and implement just the
logic to populate the `struct stat`.

[1]: https://github.molgen.mpg.de/git-mirror/glibc/blob/20003c49884422da7ffbc459cdeee768a6fee07b/sysdeps/unix/sysv/linux/generic/xstat.c#L46