Prevent random data on stat output
Some of the variables of stat struct are not initialized, thus they will
remains will unpredictable junks data from memory. This random values
could induct strange behaviours on user applications.

Even if we don't support some of the fields in the struct stat, it is
necessary to set them to some default value.