Let mercury handle its own dependencies
Searching for mercury dependencies is useless. In fact if a library has not been linked and enabled while mercury was built, we won't be able to use it.

Moreover, mercury ships with a make module that provides which links flags needs to be used. Thus there is no need to search for mercury dependencies.