Disable interception during library shutdown
If we intercept function during shutdown we could try to access the
internal file map that has been already freed.

This woun't affect performance at runtime since we are always checking
only one boolean flags on every interception as before.

On the other hand should sppedup the shutdown of the library, because we
disable the interception of glibc library call before performing the
shutdown routines.