Drop access RPC
The access RPC have been removed. The POSIX access functionality now is implemented through the stat RPC. This makes code easier to maintain and modify.

There would be a little performance degradation on the POSIX access
function because we are now retrieving also the metadata
associated to the file.

This change makes a lot easier to add symlink support