Add support for endpoint shared lookup file
For some comunication layers an out of band communication is required for putting in place the specific protocol handshake.

In those case a shared lookup file could be used to distribute deamon's endpoints information to all the clients.

This file need to resides on a shared filesystem and can be specified using the new CLI parameter `--lookup-file`.