Add CREATE_CHECK_PARENTS compilation flag
CREATE_CHECK_PARENTS controls if the existance of the parent node needs
to be checked during the creation of a child node.

Example: The check prevents the creation of file `/parent/child`
if `/parent` doesn't exists or if it is not a directory

By default CREATE_CHECK_PARENTS is enable and the check is enforced.
It can be disable to speedup creation operations by passing
`-DCREATE_CHECK_PARENTS:BOOL=OFF` as cmake parameter.